Rename Formatter => Sanitizer
