+++ classes/Commentions.php
@@ -323,7 +323,7 @@ class Commentions
                 $item['text_sanitized'] = $cachedText[$uid];
             } else if (array_key_exists('text', $item) === true) {
                 // Item has a text field, sanitize it
-                $item['text_sanitized'] = $sanitizedText[$uid] = Formatter::filter($item['text']);
+                $item['text_sanitized'] = $sanitizedText[$uid] = Sanitizer::filter($item['text']);
             }
 
             return $item;
@@ -469,7 +469,7 @@ class Commentions
      */
     protected static function cacheKey(string $pageId): string
     {
-        $suffix = Formatter::advancedFormattingAvailable() ? 'formatted' : 'escaped';
+        $suffix = Sanitizer::advancedFormattingAvailable() ? 'formatted' : 'escaped';
         return "{$pageId}-{$suffix}";
     }
 }
