+++ sections/commentions.php
@@ -68,7 +68,7 @@ return [
                 ];
             }
 
-            if (class_exists('Masterminds\\HTML5') === false || Formatter::advancedFormattingAvailable() === false) {
+            if (class_exists('Masterminds\\HTML5') === false || Sanitizer::advancedFormattingAvailable() === false) {
                 $errors[] = [
                     'id' => 'missing-dependencies',
                     'message' => t('commentions.section.error.missing-dependencies'),
